requirements.txt
flask==2.0.1
streamlit==1.2.0
kafka-python==2.0.2
pandas==1.3.3
plotly==5.3.1
requests==2.26.0